+++ src/block/SnowBlock.hpp
@@ -5,7 +5,7 @@
 struct SnowBlock : Block {
     using Block::Block;
 
-    bool isValidPosition(const BlockData& data, WorldReader &reader, const glm::vec3 &pos) override;
+    bool isValidPosition(const BlockData& data, WorldReader &reader, const glm::ivec3 &pos) override;
 
     RenderType getRenderType() const override {
         return RenderType::SnowLayer;
